qemu_devices: Workaround circular import issue
There is a circular import issue within package qemu_devices:
the qbuses and qdevices module imported each other. Currently,
we can merge these two files into one to workaround the issue.
The code structure should be re-designed in future.
